finish global sitestats router/file
